fix: rename BenchAction* -> ReleaseGroupAction*
